Split auto approval gate helpers
